Creates has-id mixin
