+++ addon/components/em-text.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
 import layout from '../templates/components/em-text';
+import hasId from '../mixins/has-id';
 
 /*
 Form Text Area
@@ -7,7 +8,7 @@ Form Text Area
 Syntax:
 {{em-text property="property name" rows=4 cols=40}}
  */
-export default Ember.Component.extend({
+export default Ember.Component.extend(hasId, {
   layout: layout,
   elementClass: null,
   htmlComponent: 'erf-html-text',
@@ -27,15 +28,5 @@ export default Ember.Component.extend({
       return 'col-sm-10';
     }
     return null;
-  }),
-
-  id: Ember.computed('cid', 'property', {
-    get: function() {
-      if (this.get('cid')) {
-        return this.get('cid');
-      } else {
-        return `${this.get('property')}-${this.elementId}`;
-      }
-    }
   })
 });
