+++ addon/templates/components/em-form.hbs
@@ -11,8 +11,6 @@
   group=(component 'form-group' form=this model=model)
 
   formGroup=(component 'em-form-group' form=this model=model)
-
-
 )}}
 {{#if submitButton}}
     {{em-form-submit form=this}}
